requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
earthengine-api==0.1.379
numpy==1.26.0
pandas==2.1.1
pydantic==2.6.0
python-dotenv==1.0.0
requests==2.31.0
httpx==0.26.0
google-generativeai==0.3.2
google-cloud-storage==2.13.0
gunicorn==21.2.0
tensorflow==2.15.0
scikit-learn==1.3.2 